consolidate logic within ui
