--- src/app/profile-logic/page.tsx
@@ -1,7 +1,0 @@
-export default function Profile() {
-  return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      Profile Page
-    </main>
-  )
-}
